-Added an I2C Altimeter (MPL3115A2)
git-svn-id: https://warg-explorer16-autopilot.googlecode.com/svn/Autopilot.X@117 0eea08a1-903a-204f-7ded-f54e91f5164d

+++ InterchipDMA.h
@@ -20,7 +20,7 @@ typedef struct _PMData {
     long double latitude;  //8 Bytes - ddd.mmmmmm
     long double longitude; //8 Bytes - ddd.mmmmmm
     float speed;    //KM/H
-    int altitude; // Meters
+    float altitude;
     int sp_Altitude; // Meters
     int heading;  //Degrees
     int sp_Heading; //Degrees
